binarisation des StopTime
git-svn-id: https://ctp-gesconf.canaltp.fr/svn/NAViTiACpp/trunk@187 a734f7ea-df10-47d3-a172-9f42111c9061

+++ source/naviMake/main.cpp
@@ -11,7 +11,7 @@ namespace pt = boost::posix_time;
 int main(int argc, char const* argv[])
 {
     pt::ptime start, end;
-    int64_t read, clean, sort, transform, save;
+    int read, clean, sort, transform, save;
 
     navimake::Data data;
     {
